+++ src/Login.js
@@ -1,7 +1,7 @@
 import React from "react";
 import {Link} from "react-router-dom";
 
-import "./css/Login.scss";
+import "./scss/Login.scss";
 import logo from "./img/appLogo.png";
 
 const Login = ({loginUser}) => {
